app/requirements.txt
flask==3.0.0
flask-pydantic==0.12.0
flask-sqlalchemy==3.1.1
gunicorn==21.2.0
mysql-connector-python==8.2.0
python-dateutil==2.8.2
redis==5.0.1
sqlalchemy==2.0.25
